gamm utils conv helpers moved to osmoutils
